+++ packages/backend/src/config/lightdashConfig.mock.ts
@@ -276,7 +276,6 @@ export const lightdashConfigMock: LightdashConfig = {
     customRoles: {
         enabled: false,
     },
-    experimentalExplorerImprovements: false,
     dashboardComments: {
         enabled: true,
     },