+++ packages/backend/src/config/parseConfig.ts
@@ -919,7 +919,7 @@ export type LightdashConfig = {
         enabled: boolean;
     };
     analyticsEmbedSecret?: string;
-    experimentalExplorerImprovements: boolean;
+
     dashboardComments: {
         enabled: boolean;
     };
@@ -1656,8 +1656,6 @@ export const parseConfig = (): LightdashConfig => {
             enabled: process.env.CUSTOM_ROLES_ENABLED === 'true',
         },
         analyticsEmbedSecret: process.env.ANALYTICS_EMBED_SECRET,
-        experimentalExplorerImprovements:
-            process.env.EXPERIMENTAL_EXPLORER_IMPROVEMENTS === 'true',
         dashboardComments: {
             enabled: process.env.DISABLE_DASHBOARD_COMMENTS !== 'true',
         },